requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
markdown-it-py==3.0.0
xhtml2pdf==0.2.15
reportlab==4.0.8
pydantic==2.5.3
python-multipart==0.0.6
aiofiles==23.2.1
